Extract out viewmodels to separate files
