+++ postdetails/src/main/java/com/neaniesoft/vermilion/postdetails/ui/PostState.kt
@@ -1,0 +1,7 @@
+package com.neaniesoft.vermilion.postdetails.ui
+
+sealed class PostState {
+    object Empty : PostState()
+    object Error : PostState()
+    data class Post(val post: com.neaniesoft.vermilion.posts.domain.entities.Post) : PostState()
+}
